+++ src/components/WeatherDisplay/WeatherDisplay.js
@@ -1,21 +1,19 @@
 import React, {useEffect} from "react";
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faSun, faSmog, faCloudRain, faSnowflake, faCloudBolt, faCloud, faMoon} from '@fortawesome/free-solid-svg-icons';
 import {useDispatch, useSelector} from "react-redux";
 import toast from "react-hot-toast";
-import {WeatherCard} from "./WeatherCard/WeatherCard";
+import {CurrentWeatherCard} from "./WeatherCards/CurrentWeatherCard";
 import styles from "./WeatherDisplay.module.css";
+import {matchCondition, thunderWeather, weatherCondition} from "../../utils/helpers";
+import {NextDayWeatherCard} from "./WeatherCards/NextDayWeatherCard";
 
 export const WeatherDisplay = () => {
 
     const {data, error} = useSelector((state) => state.weather);
     const dispatch = useDispatch();
 
-    const currentDay = data?.current;
+    const [currentDay, nextDay, theDayAfterNext] = data?.forecast?.forecastday || [];
 
-    const [currentDayInput, nextDay, theDayAfterNext] = data?.forecast?.forecastday || [];
-
-    console.log(currentDay);
+    console.log(nextDay);
 
     const condition = data?.current.condition.text.toLowerCase();
     const isDay = data?.current.is_day;
@@ -24,23 +22,7 @@ export const WeatherDisplay = () => {
 
         if (condition) {
 
-            let matchedCondition;
-
-            if (isDay === 0) {
-                matchedCondition = 'night';
-            } else if (condition.match(/rain|drizzle/i)) {
-                matchedCondition = 'rainy';
-            } else if (condition.match(/snow|sleet|ice|blizzard/i)) {
-                matchedCondition = 'snowy';
-            } else if (condition.match(/mist|fog/i)) {
-                matchedCondition = 'misty';
-            } else if (condition.match(/cloudy|overcast/i)) {
-                matchedCondition = 'cloudy';
-            } else {
-                matchedCondition = 'sunny';
-            }
-
-            dispatch({type: 'COLOR_CHANGE', payload: matchedCondition});
+            dispatch({type: 'COLOR_CHANGE', payload: matchCondition(condition,isDay)});
         }
 
     }, [condition, isDay]);
@@ -52,44 +34,6 @@ export const WeatherDisplay = () => {
             </>
         );
     }
-    const thunderWeather = (condition) => {
-
-        if (condition.includes("thunder")) {
-            return (
-                <FontAwesomeIcon icon={faCloudBolt} shake size="2xl"/>
-            )
-        }
-
-    }
-
-    const weatherCondition = (condition) => {
-
-        if (condition.match(/rain|drizzle/i)) {
-            return (
-                <FontAwesomeIcon icon={faCloudRain} bounce size="2xl"/>
-            )
-        } else if (condition.match(/snow|sleet|ice|blizzard/i)) {
-            return (
-                <FontAwesomeIcon icon={faSnowflake} spin size="2xl"/>
-            )
-        } else if (condition.match(/mist|fog/i)) {
-            return (
-                <FontAwesomeIcon icon={faSmog} fade size="2xl"/>
-            )
-        } else if (condition.match(/cloudy|overcast/i)) {
-            return (
-                <FontAwesomeIcon icon={faCloud} fade size="2xl"/>
-            )
-        } else if (isDay === 0) {
-            return (
-                <FontAwesomeIcon icon={faMoon} flip size="2xl"/>
-            )
-        } else {
-            return (
-                <FontAwesomeIcon icon={faSun} beat size="2xl"/>
-            )
-        }
-    }
 
     return (
         <>
@@ -97,17 +41,17 @@ export const WeatherDisplay = () => {
                 <section className={styles.weather}>
                     <h2>{data?.location.name}</h2>
                     <article className={styles.weather__cards}>
+                        <article className={styles.weather__cards_nextDays}>
+                            <NextDayWeatherCard data={nextDay} thunderWeather={thunderWeather}
+                                                weatherCondition={weatherCondition}/>
+                        </article>
                         <article className={styles.weather__cards_currentDay}>
-                            <WeatherCard data={data} thunderWeather={thunderWeather}
-                                         weatherCondition={weatherCondition} condition={data?.current.condition.text.toLowerCase()}/>
+                            <CurrentWeatherCard data={data} thunderWeather={thunderWeather}
+                                                weatherCondition={weatherCondition} />
                         </article>
                         <article className={styles.weather__cards_nextDays}>
-                            <WeatherCard data={data} thunderWeather={thunderWeather}
-                                         weatherCondition={weatherCondition}/>
-                        </article>
-                        <article className={styles.weather__cards_nextDays}>
-                            <WeatherCard data={data} thunderWeather={thunderWeather}
-                                         weatherCondition={weatherCondition}/>
+                            <NextDayWeatherCard data={theDayAfterNext} thunderWeather={thunderWeather}
+                                                weatherCondition={weatherCondition}/>
                         </article>
                     </article>
                 </section>)}
